api/requirements.txt
anyio==3.5.0
asgiref==3.5.0
boto3==1.20.54
botocore==1.23.54
click==8.0.3
fastapi==0.73.0
geojson==2.5.0
h11==0.13.0
idna==3.3
jmespath==0.10.0
pydantic==1.9.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python-multipart==0.0.5
s3transfer==0.5.1
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.1.0
urllib3==1.26.8
uvicorn==0.17.4
